Finished app

+++ red_tape/lib/Models/invention.dart
@@ -5,12 +5,12 @@ import 'user.dart';
 class Invention {
   String name;
   List<User> inventors;
-  int progress;
+  double progress;
   List<User> investors;
   String category;
   AssetImage image;
 
-  Invention(String name, List<User> inventors, List<User> investors, int progress, String category, AssetImage image) {
+  Invention(String name, List<User> inventors, List<User> investors, double progress, String category, AssetImage image) {
     this.name = name;
     this.inventors = inventors;
     this.investors = investors;
